+++ R/droogte_transformatie_KNMI14.R
@@ -1,35 +1,23 @@
 #' Calculation of Makkink evaporation
 #' @description Function reads transormed mean temperature and transformed global radiation
 #' and calculates the Makkink evaporation for 'future time series' that match a certain climate
-#' @inheritParams temperatuur_transformatie_KNMI14
+#' @inheritParams TransformTemp
 #' @param ifile_tg   Name of the input file for temperature
 #' @param ifile_rsds Name of the input file for radiation
-#' @param regio.file     this (optional) argument provides the name of an ASCII file that relates the stations to
-#'                a particular region. First column is station id and second column region
-#'                KNMI14 distinguishes following regions:
-#'                <NLD> Nederland            [DEFAULT]
-#'                <NWN> Noordwest Nederland
-#'                <ZWN> Zuidwest Nederland
-#'                <NON> Noordoost Nederland
-#'                <MON> Middenoost Nederland
-#'                <ZON> Zuidoost Nederland
-#'
 #' @export
-droogte_berekening_KNMI14 <- function(ifile_tg, ifile_rsds,
-                                      ofile=NA,
-                                      sc,
-                                      p=NA,
-                                      regio.file = NA) {
+TransformEvap <- function(ifile_tg, ifile_rsds,
+                          ofile = NA,
+                          scenario,
+                          horizon = 2030,
+                          regio.file = NA) {
 
-  version <- ReturnPackageVersion()
-  # CONSTANTS AND FUNCTIONS ###############################################################################
+  version <- ReturnPackageVersion("evmk")
 
-  CheckPeriod(p)
+  CheckPeriod(horizon)
 
-  ## Need to add an IF for delta.files of rsds & tg = need to be for the same p and the same sc
-
-  rsds_input <- straling_transformatie_KNMI14(ifile = ifile_rsds, sc=sc, p=p)
-  tg_input   <- temperatuur_transformatie_KNMI14(ifile = ifile_tg, var="tg", sc=sc, p=p, regio.file = regio.file)
+  rsds_input <- TransformRadiation(ifile = ifile_rsds, scenario=scenario, horizon = horizon)
+  tg_input   <- TransformTemp(ifile = ifile_tg, var="tg", scenario=scenario,
+                              horizon = horizon, regio.file = regio.file)
 
   rsds <- rsds_input[-(1:5)]
   tg   <- tg_input[-(1:5)]
@@ -52,7 +40,7 @@ droogte_berekening_KNMI14 <- function(ifile_tg, ifile_rsds,
   result[, V1 := as.integer(V1)]
 
   if (!is.na(ofile)) {
-    WriteOutput("evmk", ofile, version, sc, p, H.comments, result)
+    WriteOutput("evmk", ofile, version, scenario, horizon, H.comments, result)
   }
 
   flog.debug("Evaporation calculation ended successfully!")